Add ImageMagick/RMagick for SEOWorks Image Sizes
SEOWorks seems to have a bug where occasionally images are returned
with valid URLs but with `null` for both the width and height dimensions.

Here we add ImageMagick and the RMagick gem for its Ruby bindings to
let us download the image and get hte dimentions ourselves if necessary.

If and when we hear back from SEOWorks this should be reverted/removed,
since it adds a huge amount of extra stuff that I'd prefer not to have
hanging around.

Rebuild all containers after upgrading