Add ActiveJob to warm caches for articles
This affects all but is only tested for the moment on SNWorks.

This adds a job to warm the caches for articles. Since the round trip
can be very heavy (especially for SNWorks, where there are 20+ requests
per single api call), this caches the responses. For the moment every 30
seconds for testing.