Rearrange paid native advertising so its not top
The original organization that this integration was created for does
paid native advertising. The only way to determine which is paid however
is to look at the author, which in the original case is 'Scholarship
Media'. This will rearrange the articles so that the paid ad is never at
the top.