Set SNWorks to return only published articles
For some very dumb quirk of development (probably tech debt)
SNWorks is apparently a two tiered system, where the backend API
is not synced completely with the front end. This is news to me,
and greatly messes some stuff up. Specifically if, for instance,
an article is set to publish later, the search results will still
find it. This is bad.

So instead we now check if `published_at` exists, and if it's after
right now. We also retriev 10 items now and choose the first ten that
return something.